refactor: 클린 아키텍쳐를 위해 파일 구조 변경
