+++ app/src/main/AndroidManifest.xml
@@ -17,11 +17,11 @@
         android:theme="@style/Theme.Map"
         tools:targetApi="31">
         <activity
-            android:name=".SearchActivity"
+            android:name=".presentation.SearchActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustResize"/>
         <activity
-            android:name=".MapActivity"
+            android:name=".presentation.MapActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustResize">
             <intent-filter>
