+++ app/src/main/java/campus/tech/kakao/map/data/repository/HistoryRepositoryImpl.kt
@@ -1,0 +1,18 @@
+package campus.tech.kakao.map.data.repository
+
+import campus.tech.kakao.map.domain.model.Location
+import campus.tech.kakao.map.domain.repository.HistoryRepository
+
+class HistoryRepositoryImpl : HistoryRepository {
+    override fun insertHistory(newHistory: Location) {
+        TODO("Not yet implemented")
+    }
+
+    override fun deleteHistory(oldHistory: Location) {
+        TODO("Not yet implemented")
+    }
+
+    override fun getAllHistory(): List<Location> {
+        TODO("Not yet implemented")
+    }
+}
