+++ app/src/main/java/campus/tech/kakao/map/data/repository/ResultRepositoryImpl.kt
@@ -1,0 +1,10 @@
+package campus.tech.kakao.map.data.repository
+
+import campus.tech.kakao.map.domain.model.Location
+import campus.tech.kakao.map.domain.repository.ResultRepository
+
+class ResultRepositoryImpl : ResultRepository {
+    override fun getAllResult(): List<Location> {
+        TODO("Not yet implemented")
+    }
+}
