+++ app/src/main/java/campus/tech/kakao/map/data/source/MapDbHelper.kt
@@ -1,8 +1,9 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.data.source
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
+import campus.tech.kakao.map.MapContract
 
 class MapDbHelper(mContext: Context) :
     SQLiteOpenHelper(mContext, DATABASE_NAME, null, DATABASE_VERSION) {
