+++ app/src/main/java/campus/tech/kakao/map/data/source/RetrofitService.kt
@@ -1,5 +1,6 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.data.source
 
+import campus.tech.kakao.map.data.ServerResult
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Header
