+++ app/src/main/java/campus/tech/kakao/map/domain/model/Location.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.domain.model
 
 import java.io.Serializable
