+++ app/src/main/java/campus/tech/kakao/map/domain/repository/HistoryRepository.kt
@@ -1,0 +1,9 @@
+package campus.tech.kakao.map.domain.repository
+
+import campus.tech.kakao.map.domain.model.Location
+
+interface HistoryRepository {
+    fun insertHistory(newHistory: Location)
+    fun deleteHistory(oldHistory: Location)
+    fun getAllHistory(): List<Location>
+}
