+++ app/src/main/java/campus/tech/kakao/map/domain/repository/ResultRepository.kt
@@ -1,0 +1,7 @@
+package campus.tech.kakao.map.domain.repository
+
+import campus.tech.kakao.map.domain.model.Location
+
+interface ResultRepository {
+    fun getAllResult(): List<Location>
+}
