+++ app/src/main/java/campus/tech/kakao/map/presentation/ResultRecyclerAdapter.kt
@@ -1,12 +1,14 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.presentation
 
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
+import campus.tech.kakao.map.DatabaseListener
+import campus.tech.kakao.map.domain.model.Location
+import campus.tech.kakao.map.R
 
 class ResultRecyclerAdapter(
     var searchResult: List<Location>,
